CORE-1510 Renamed FineGrained RSpace to just RSpace.
