+++ models/src/test/scala/coop/rchain/models/rholang/SortTest.scala
@@ -52,7 +52,7 @@ class ScoredTermSpec extends FlatSpec with PropertyChecks with Matchers {
     unsortedTerms.sorted should be(sortedTerms)
   }
   it should "sort so that unequal terms have unequal scores and the other way around" in {
-    def checkScoreEquality[A: Sortable: Arbitrary]: Unit = {
+    def checkScoreEquality[A: Sortable: Arbitrary]: Unit =
       // ScalaCheck generates similar A-s in subsequent calls which
       // we need to hit the case where `x == y` more often
       forAll(Gen.listOfN(5, arbitrary[A])) { as: List[A] =>
@@ -63,7 +63,6 @@ class ScoredTermSpec extends FlatSpec with PropertyChecks with Matchers {
             assert(sort(x).score == sort(y).score)
         }
       }
-    }
     checkScoreEquality[Bundle]
     checkScoreEquality[Connective]
     checkScoreEquality[Expr]
